middleware: integrating permissions and graphs
This change adds permissions check to the graph repo. Currently it only
works with the twitter (2009) and the facebook (caltech36) graphs with
neighbor lookups. The graph repo calls the middleware, the middleware
looks up the pointers in the shared memory, sets up permissions and then
returns a pointer to the graph.

Signed-off-by: Kaustav Goswami <kggoswami@ucdavis.edu>